+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `callToActionLinkTarget` and `linkTarget` props to the `InfoCard` component.
 
 ## [3.113.0] - 2020-04-30
 ### Added